--- src/core/drivers/driver.jl
@@ -1,4 +1,0 @@
-
-
-# The driver layer would transform the inputs into actual I/O workloads 
-# for the corresponding transport backend
