+++ src/hello/hello.jl
@@ -1,10 +1,10 @@
-using MPI
+import MPI
 import MLStyle: @match
 
 function hello(config::Config)
-    worldrank = getmpiworldrank()
-    worldsize = getmpiworldsize()
-    # TODO: Should this be a command-line argument for the hello sub-command
+    worldrank = MPI.Comm_rank(MPI.COMM_WORLD)
+    worldsize = MPI.Comm_size(MPI.COMM_WORLD)
+    # TODO: Should this be a command-line argument for the hello sub-command?
     backend = config[:datasets][1][:io_backend]
     basename = "hello"
     data = "Hello world, I am rank $worldrank of $worldsize\n"
