+++ test/runtests.jl
@@ -11,9 +11,12 @@ macro test_exists_and_rm(fn)
 end
 
 # unit tests
+include("unit/helper/test_ratios.jl")
+
 include("unit/core/test_Ctrl.jl")
 include("unit/core/test_inputs.jl")
-include("unit/core/test_hello.jl")
+
+include("unit/hello/test_hello.jl")
 
 # functional tests
 include("functional/test_main.jl")
